errors_casting

+++ Includes/cub.h
@@ -44,10 +44,10 @@ typedef struct s_data
 	int result_lenght;
 	int result_x_length;
 
-	double px_y;
-	double py_y;
-	double px_x;
-	double py_x;
+	int px_y;
+	int py_y;
+	int px_x;
+	int py_x;
 }	t_data;
 
 char	*get_next_line(int fd);
